add custom exception - NotFoundException

+++ PetHotel.Application/Hotel/Command/DeleteHotel/DeleteHotelCommandHandler.cs
@@ -1,16 +1,21 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using PetHotel.Domain.Exceptions;
 using PetHotel.Domain.Repositories;
 
 namespace PetHotel.Application.Hotel.Command.DeleteHotel;
 
-public class DeleteHotelCommandHandler(ILogger<DeleteHotelCommandHandler> logger, IMapper mapper, IHotelRepository hotelRepository) : IRequestHandler<DeleteHotelCommand>
+public class DeleteHotelCommandHandler(ILogger<DeleteHotelCommandHandler> logger, IHotelRepository hotelRepository) : IRequestHandler<DeleteHotelCommand>
 {
     public async Task Handle(DeleteHotelCommand request, CancellationToken cancellationToken)
     {
         logger.LogInformation("Deleting hotel with id {Id}", request.Id);
         var hotel = await hotelRepository.GetHotelByIdAsync(request.Id);
+        if (hotel is null)
+        {
+            throw new NotFoundException(nameof(hotel), request.Id.ToString());
+        }
         await hotelRepository.DeleteHotel(hotel);
     }
 }
